Tighten the export predicate and dedup validate() within a pass (review)
Three findings from review, all accepted.

`exportProducedPayload(at:)` now requires a regular file. A directory reports
a nonzero size of its own, so the size test alone would have passed one;
`exportPayload(id:)` only ever exports `ExportType.file`, so a directory at
that path is not a payload of ours.

Both "the payload is already at the destination" fallbacks in
`exportPayload(reference:fileName:)` — the new one on a failed export and the
long-standing one in the move's catch — now apply that same predicate instead
of a bare `fileExists`. An empty file at the destination was being returned as
a successful export, which is precisely the silent hole #388 is about: the
attachment resolves to a URL, so `validate()` sees content and records nothing.
The #449 race is unaffected, since the winner of that race leaves a complete
file behind.

`exportPayloadData` registers its temp-file cleanup before the payload check
rather than after, so a half-made file an unsuccessful export left behind is
removed too. The nil-unwrap and the content check stay in one guard: splitting
them to make room for the `defer` would restore the standalone dead branch
this PR exists to remove.

`Summary.validate()` now dedups within a single pass as well as across calls.
The sets are seeded from what is already recorded and grown as the pass
records, so one bundle passed twice yields one entry rather than two the
reader cannot tell apart. This applies to attachments as well as logs — the
attachment loop had the same snapshot-only behaviour before this PR.
`testTwoRunsLosingTheSameLogAreRecordedOnce` pins it, and
`testExportedPayloadIsOnlyAcceptedWhenItHasBytes` grows a directory case.

The #386 tests move to `LogFaultReportingTests` alongside the #388 ones,
which keeps `FaultReportingTests` inside SwiftLint's type-body limit and
leaves one file per issue.
