+++ Tests/XCTestHTMLReportTests/LogFaultReportingTests.swift
@@ -1,0 +1,173 @@
+//
+//  LogFaultReportingTests.swift
+//
+//  A run's log must not go missing quietly (#386). `FaultReportingTests`
+//  covers the attachment half of the same post-condition.
+//
+
+import XCTest
+@testable import XCTestHTMLReportCore
+
+final class LogFaultReportingTests: XCTestCase {
+    /// A summary built entirely from constants, so the only thing that can
+    /// leave a run's log empty is the reference failing to resolve.
+    ///
+    /// `StubPayloadProvider` records no faults of its own, which is the point:
+    /// anything a test finds on the collector afterwards came from the
+    /// post-condition, not from a call site.
+    private func syntheticSummary(
+        logReference: String?,
+        runCount: Int = 1,
+        collector: FaultCollector
+    ) -> Summary {
+        Summary(
+            parsedRuns: (0 ..< runCount).map { _ in
+                ParsedRun(
+                    destination: ParsedDestination(
+                        displayName: "Synthetic Device",
+                        deviceIdentifier: "00000000-0000-0000-0000-000000000000",
+                        modelName: "Synthetic Model",
+                        operatingSystemVersion: "1.0"
+                    ),
+                    logReference: logReference,
+                    testables: []
+                )
+            },
+            payloads: SyntheticResult.payloads,
+            renderingMode: .linking,
+            downsizeImagesEnabled: false,
+            downsizeScaleFactor: 0.5,
+            faultCollector: collector,
+            bundleNames: ["Synthetic"]
+        )
+    }
+
+    func testRunWithoutLogReferenceIsNotAResolutionFailure() {
+        // A run can legitimately carry no log reference at all. Its log
+        // content is `.none` by construction, not through degradation — the
+        // same rule that keeps a payload-less attachment out of the fault list
+        // (#387).
+        let collector = FaultCollector()
+        let summary = syntheticSummary(logReference: nil, collector: collector)
+
+        summary.validate()
+
+        XCTAssertEqual(
+            summary.faults, [],
+            "A run that never had a log reference must not be reported as degradation"
+        )
+    }
+
+    func testRunWhoseLogCannotBeResolvedIsRecordedAsFault() {
+        // The counterpart: a reference existed and resolved to nothing, so the
+        // report ships without that run's log. Nothing recorded that before
+        // #386, and the run exited 0.
+        let collector = FaultCollector()
+        let summary = syntheticSummary(logReference: "no-such-log", collector: collector)
+
+        XCTAssertEqual(
+            collector.faults, [],
+            "The provider is expected to fail silently here; otherwise the "
+                + "assertion below would not be testing the post-condition"
+        )
+
+        summary.validate()
+
+        XCTAssertEqual(
+            summary.faults.map(\.kind), [.unresolvedLog],
+            "A log reference that resolved to no content must be recorded as degradation"
+        )
+
+        let afterFirst = summary.faults
+        summary.validate()
+        XCTAssertEqual(
+            summary.faults, afterFirst,
+            "validate() must not re-record the log fault on a second call"
+        )
+    }
+
+    func testTwoRunsLosingTheSameLogAreRecordedOnce() {
+        // Passing one bundle twice produces two runs whose fault details are
+        // identical, since the detail names the destination. Two entries the
+        // reader cannot tell apart is noise, so dedup applies within a single
+        // pass and not only across repeated calls.
+        let collector = FaultCollector()
+        let summary = syntheticSummary(
+            logReference: "no-such-log", runCount: 2, collector: collector
+        )
+
+        summary.validate()
+
+        XCTAssertEqual(
+            summary.faults.map(\.kind), [.unresolvedLog],
+            "One indistinguishable degradation is one fault"
+        )
+    }
+
+    func testRunWithAResolvableLogIsNotAFault() {
+        let collector = FaultCollector()
+        let summary = syntheticSummary(
+            logReference: SyntheticResult.logReference, collector: collector
+        )
+        _ = summary.generatedHtmlReport()
+
+        summary.validate()
+
+        XCTAssertEqual(summary.faults, [], "A log that resolved must not be flagged")
+    }
+
+    /// Copies `SanityResults.xcresult` and deletes the object its log
+    /// reference names, leaving a bundle whose log reference survives parsing
+    /// and then resolves to nothing.
+    private func makeBundleWithUnreadableLog() throws -> URL {
+        let source = try XCTUnwrap(
+            Bundle.testBundle.url(forResource: "SanityResults", withExtension: "xcresult")
+        )
+        let scratch = URL(fileURLWithPath: NSTemporaryDirectory())
+            .appendingPathComponent("xchr-broken-log-\(UUID().uuidString)")
+        let bundle = scratch.appendingPathComponent("SanityResults.xcresult")
+        try FileManager.default.createDirectory(at: scratch, withIntermediateDirectories: true)
+        addTeardownBlock { try? FileManager.default.removeItem(at: scratch) }
+        try FileManager.default.copyItem(at: source, to: bundle)
+
+        let reference = try XCTUnwrap(
+            LegacyResultReader(file: ResultFile(url: bundle, faultCollector: FaultCollector()))
+                .read()?.runs.first?.logReference,
+            "Fixture is expected to carry a log reference"
+        )
+        // The bundle is a content-addressed store: `Data/data.<id>` holds the
+        // object `<id>` names, and both backends resolve the run log through
+        // it. Removing it is the reachable stand-in for the decode failure
+        // #386 is about.
+        try FileManager.default.removeItem(
+            at: bundle.appendingPathComponent("Data/data.\(reference)")
+        )
+        return bundle
+    }
+
+    func testUnreadableLogExitsNonZeroIdenticallyOnBothBackends() throws {
+        // Constructing the bundle needs the legacy reader to name the log
+        // object; skip with the rest of the legacy suite once it is gone.
+        if case .legacyUnavailable = ResultBackend.legacy.resolve() {
+            throw XCTSkip("Toolchain has no legacy commands; the legacy leg cannot run.")
+        }
+        let bundle = try makeBundleWithUnreadableLog()
+
+        for reader in ["legacy", "modern"] {
+            let (status, stdOut, stdErr) = try xchtmlreportCmd(
+                args: ["--result-reader", reader, bundle.path]
+            )
+            let combined = (stdOut ?? "") + (stdErr ?? "")
+
+            XCTAssertEqual(
+                status, 3,
+                "\(reader): a report shipped without its run log must not exit 0.\n\(combined)"
+            )
+            // The cause, whose detail is backend-internal, and the symptom,
+            // whose detail is not: both backends name the same run the same
+            // way, so structurally equal inputs degrade identically.
+            try XCTAssertContains(combined, "logExportFailed")
+            try XCTAssertContains(combined, "unresolvedLog: log for run ")
+        }
+    }
+}
